refactored mapping inheritance
